+++ api/users.js
@@ -1,0 +1,2 @@
+import { usersHandler } from '../server/auth-handlers.js';
+export default usersHandler();